hackathon-submissions/s3-data-migration-cstor/webapp/pyproject.toml
[tool.poetry]
name = "cstor_dashboard"
version = "0.1.0"
description = ""
authors = ["olwee <hello@codewitholi.com>"]
packages = [
  { include = "cstor", from = "src" }
]

[[tool.poetry.source]]
name = "gitlab"
url = "https://gitlab.com/api/v4/projects/35836634/packages/pypi/simple"
secondary = true

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^1.9.0"
pytest-mock = "^3.7.0"
Django = "^4.0.4"
django-restframework = "^0.0.1"
django-cors-headers = "^3.12.0"
djoser = "^2.1.0"
cstor-s3 = "0.4.4"
channels = "^3.0.4"
channels-redis = "^3.4.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
coverage = "^6.3.2"
black = "^22.3.0"
isort = "^5.10.1"
pylint = "^2.13.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pylint.MASTER]
fail-under = "10"
output-format = "parseable"
msg-template = "{path}:{line}: [{msg_id}({symbol}), {obj}] {msg}"
